+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/basic/Logarithm.kt
@@ -1,7 +1,7 @@
 package io.github.kapkekes.operations.basic
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 import io.github.kapkekes.complex.extensions.log
 
 /** Returns the logarithm of the first number by base of the second. */
