+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/basic/Power.kt
@@ -1,7 +1,7 @@
 package io.github.kapkekes.operations.basic
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 
 /** Returns the power of the first number to the second. */
 object Power : Operation {
