+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/flow/Exit.kt
@@ -1,7 +1,7 @@
 package io.github.kapkekes.operations.flow
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 
 /** Throws [RuntimeException]. */
 object Exit : Operation {
