+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/logic/Greater.kt
@@ -1,7 +1,7 @@
 package io.github.kapkekes.operations.logic
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 
 /**
  * Returns [True], if the numbers are real and the first number is greater than the second.
