+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/trigonometry/Cosine.kt
@@ -1,7 +1,7 @@
 package io.github.kapkekes.operations.trigonometry
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 import io.github.kapkekes.complex.extensions.cos
 
 /** Returns the cosine of the given number. */
